% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{comp.gene.transc.levels}
\alias{comp.gene.transc.levels}
\title{Compile the transcription levels for the coding regions.}
\usage{
comp.gene.transc.levels(genes, cd)
}
\arguments{
\item{genes}{An annotation data table.}

\item{cd}{A data table representing the coverage depth for a given RNA-seq expression profile.}
}
\description{
Internal function to quantify the expression at CDS-level.
}
\author{
Vittorio Fortino
comp.gene.transc.levels()
}
\keyword{internal}

