% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{comp.igr.transc.levels}
\alias{comp.igr.transc.levels}
\title{Compile the transcription levels for the intergenic regions.}
\usage{
comp.igr.transc.levels(genes, igrs, cd, transcCDSs, str = "+")
}
\arguments{
\item{genes}{Data table containing gene annotations.}

\item{cd}{Data table containing the coverage depth of an RNA-seq expression profile(s).}

\item{transcCDSs}{Transcription levels for the coding regions.}

\item{str}{Charcater value indicating the strand. Defauls value is "+".}

\item{genes}{Data tbale containing intergenic region annotations.}
}
\description{
Internal function to quantify expression at the intergenic-level.
}
\author{
Vittorio Fortino
comp.igr.transc.levels()
}
\keyword{internal}

