 \name{ct1}
 \docType{data}
 \alias{ct1}
 \title{Coverage Vector 1}
 \description{
    An example of data frame containing the transcriptome RNA-seq profile of a prokaryotic organism studied under a specific experimental condition.
    It contains the coverage vectors for both strands: fwd (coverage depth on the forward strand) and rev (coverage depth on reverse strand). 
 }
 \value{
    Return a data frame having two columns: fwd (coverage depth on the forward strand) and rev (coverage depth on reverse strand).
 }
 \usage{data(ct1)}
 \format{A data frame.}