% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{get.intergenic.regions}
\alias{get.intergenic.regions}
\title{Build a data table containing generic information on intergenic regions.}
\usage{
get.intergenic.regions(genes, str = "+")
}
\arguments{
\item{genes}{An annotation data table.}

\item{str}{A given strand. Defaults to "+".}
}
\description{
Internal function to build a data table containing information of the intergenic regions on a given strand.
}
\author{
Vittorio Fortino
get.intergenic.regions()
}
\keyword{internal}

