% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{get.NCBI.seq}
\alias{get.NCBI.seq}
\title{Find and get a genome sequence by specifyng the accession number. Read gene annotations in GFF format}
\usage{
get.NCBI.seq(accession, list.dbs = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{Indicate whether information about the process should be reported. Defaults to TRUE.}

\item{gff_file}{A full local path indicating the GFF file to load.}
}
\value{
Annotation data table.
}
\description{
Find and get a genome sequence by specifyng the accession number. Read gene annotations in GFF format
}
\author{
Vittorio Fortino
}
\keyword{internal}

