% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{join.genes.and.operons}
\alias{join.genes.and.operons}
\title{Join gene(s) and operon(s) annotations.}
\usage{
join.genes.and.operons(genes, ops)
}
\arguments{
\item{genes}{An annotation data table.}

\item{ops}{An operon data table.}
}
\value{
Operon data table.
}
\description{
Internal function to merge the annotations with DOOR data.
}
\author{
Vittorio Fortino
join.genes.and.operons()
}
\keyword{internal}

