% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{pred.operon.status}
\alias{pred.operon.status}
\title{Predict operon status of gene pairs (e.g., POPs, DOPs,...).}
\usage{
pred.operon.status(list.cls, gene.pairs, type = "", cons = 1,
  verbose = TRUE, ...)
}
\arguments{
\item{gene.pairs}{Gene pairs.}

\item{type}{Character vector indicating the type of gene pairs (POPs or DOPs). See \code{select.pops} and \code{select.ops.indoor}.}

\item{cons}{Numeric values indicating the minimum number of postive votes to classify a gen pair as operon pai. See \code{select.pops} and \code{select.ops.indoor}.}

\item{verbose}{Default logical value is TRUE.}

\item{cls}{List of classifiers. See \code{train.rfs}.}
}
\description{
Internal function that uses the trained operon classifier to predict the operon status of POPs and DOPs.
}
\author{
Vittorio Fortino
}
\keyword{internal}

