% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{qcut}
\alias{qcut}
\title{Determine cutoff values for a given RNA-seq expression profile.}
\usage{
qcut(data)
}
\arguments{
\item{data}{A coverage-depth table.}
}
\description{
Internal function to estimate the cutoff values to distinguish low expressed RNA-seq data from high expression data.
}
\author{
Vittorio Fortino
qcut()
}
\keyword{internal}

