% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{read.annot.from.gff}
\alias{read.annot.from.gff}
\title{Read a GFF file from NCBI and return a GRanges object.}
\usage{
read.annot.from.gff(gff.file, locus.tags = TRUE, nrows = -1,
  verbose = TRUE)
}
\arguments{
\item{gff.file}{a GFF file}

\item{locus.tags}{only return genes with locus tags. Defaults to TRUE.}

\item{nrows}{number of rows to read.}

\item{verbose}{Indicate whether information about the process should be reported. Defaults to TRUE.}
}
\value{
GRanges with 4 elementMetadata columns: locus, proetin id, gene id, feature, description and gene name. 
        If all rows are returned (locus.tags=FALSE), then score, phase and tags are included.
}
\description{
Read a GFF file from NCBI and return a GRanges object.
}
\author{
Vittorio Fortino
read.annot.from.gff()
}

