% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{read.door.annotations}
\alias{read.door.annotations}
\title{Read the operon(s) data file downloaded from http://csbl.bmb.uga.edu/DOOR/displayspecies.php}
\usage{
read.door.annotations(opr_file, verbose = TRUE)
}
\arguments{
\item{opr_file}{A full local path indicating the .opr file to load.}

\item{verbose}{Indicate whether information about the process should be reported. Defaults to TRUE.}
}
\value{
Operon data table.
}
\description{
Internal function to read the operon(s) data file downloaded from http://csbl.bmb.uga.edu/DOOR/displayspecies.php
}
\author{
Vittorio Fortino
read.door.annotations()
}
\keyword{internal}

