% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{select.ops.indoor}
\alias{select.ops.indoor}
\title{Define the set of OPs annotated in DOOR.}
\usage{
select.ops.indoor(genes.and.operons, transcCDSs, transcIGRs.pos, transcIGRs.neg,
  wseq, verbose = TRUE, ...)
}
\arguments{
\item{genes.and.operons}{Data table merging gene(s) and operon(s) annotations. See \code{join.genes.and.operons}.}

\item{transcCDSs}{Transcription levels for the coding regions. See \code{comp.gene.transc.levels}.}

\item{transcIGRs.pos}{Transcription levels for the intergenic regions (forard strand). See \code{comp.igr.transc.levels}.}

\item{transcIGRs.neg}{Transcription levels for the intergenic regions (reverse strand). See \code{comp.igr.transc.levels}.}

\item{wseq}{Sequence genome.}

\item{verbose}{Default logical value is TRUE.}
}
\description{
Build a data table containing the transcriptomic/genomic feature values of operon pairs annotated in DOOR.
The operon classifier trained/validated on the OPs and NOPs is used to predict the operon status of these gene pairs.
}
\author{
Vittorio Fortino
select.ops.indoor()
}
\keyword{internal}

