% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{sum.conf.matrix}
\alias{sum.conf.matrix}
\title{Provide classification metrics.}
\usage{
\method{sum}{conf.matrix}(cm, ...)
}
\arguments{
\item{cm}{confusion matrix.}
}
\description{
Internal function to compile common classification metrics on a give confusion matrix.
}
\author{
Vittorio Fortino
}
\keyword{internal}

