% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{test.corr}
\alias{test.corr}
\title{Statistical test to find a putative transcription start (or end) point.}
\usage{
test.corr(x, sharpInc = 1, cutCorr = 0.7, cutPVal = 1e-07, flag = 0)
}
\arguments{
\item{x}{Segment of coverage depth.}

\item{sharpInc}{Default value is 1.}

\item{cutCorr}{Cutoff value for the correlation coefficient. Default values is 0.7.}

\item{cutPVal}{Cutoff value for the p-value. Default values is 0.0000001.}

\item{flag}{Indicate weather the correlation test is for a start- or end-point in transcription. Default values is 0 (start-point).}
}
\description{
Test the correlation coefficient between a segment of coverage depth and a vector of 100 integers modeling 
a simple shape of sharp increases (or decreases) in transcription: x =???[0..0,1..1] (or x =???[1..1,0..0]).
}
\author{
Vittorio Fortino
test.corr()
}
\keyword{internal}

