% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONDOP.R
\name{validate.cls}
\alias{validate.cls}
\title{Validate the classification models.}
\usage{
validate.cls(data, class, models, runs = 5, kf = 5, verbose = TRUE)
}
\arguments{
\item{data}{Training/test data. See \code{select.ops} and \code{select.nops}.}

\item{class}{Vector of the class labels.}

\item{runs}{Number of bootstraps to be used.}

\item{kf}{Number of folds for the cross-validation.}
}
\description{
Internal function to validate all the classifiers to be used in distinguishing operon pairs (OPs) from non-operon pairs (NOPs)
on a given RNA-seq expression profile.
}
\author{
Vittorio Fortino
}
\keyword{internal}

